requirements.txt
requests
python-dotenv
playwright
Jinja2